chore: decouple dynamic dispatch boiler plate in transaction
