+++ keystore/src/entities/platform/generic/mls/e2ei_acme_ca.rs
@@ -49,8 +49,8 @@ impl EntityBase for E2eiAcmeCA {
         MissingKeyErrorKind::E2eiAcmeCA
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::E2eiAcmeCA(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::E2eiAcmeCA(self)
     }
 
     async fn find_all(conn: &mut Self::ConnectionType, params: EntityFindParams) -> CryptoKeystoreResult<Vec<Self>> {
