+++ keystore/src/entities/platform/generic/mls/e2ei_crl.rs
@@ -37,8 +37,8 @@ impl EntityBase for E2eiCrl {
         MissingKeyErrorKind::E2eiCrl
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::E2eiCrl(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::E2eiCrl(self)
     }
 
     async fn find_all(
