+++ keystore/src/entities/platform/generic/mls/encryption_keypair.rs
@@ -43,8 +43,8 @@ impl EntityBase for MlsEncryptionKeyPair {
         MissingKeyErrorKind::MlsEncryptionKeyPair
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::EncryptionKeyPair(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::EncryptionKeyPair(self)
     }
 
     async fn find_all(
