+++ keystore/src/entities/platform/generic/mls/epoch_encryption_keypair.rs
@@ -44,8 +44,8 @@ impl EntityBase for MlsEpochEncryptionKeyPair {
         MissingKeyErrorKind::MlsEpochEncryptionKeyPair
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::EpochEncryptionKeyPair(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::EpochEncryptionKeyPair(self)
     }
 
     async fn find_all(
