+++ keystore/src/entities/platform/generic/mls/keypackage.rs
@@ -44,8 +44,8 @@ impl EntityBase for MlsKeyPackage {
         MissingKeyErrorKind::MlsKeyPackageBundle
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::KeyPackage(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::KeyPackage(self)
     }
 
     async fn find_all(
