+++ keystore/src/entities/platform/generic/mls/pending_message.rs
@@ -43,8 +43,8 @@ impl EntityBase for MlsPendingMessage {
         MissingKeyErrorKind::MlsPendingMessages
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::MlsPendingMessage(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::MlsPendingMessage(self)
     }
 
     async fn find_one(
