+++ keystore/src/entities/platform/generic/proteus/identity.rs
@@ -41,8 +41,8 @@ impl EntityBase for ProteusIdentity {
         MissingKeyErrorKind::ProteusIdentity
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::ProteusIdentity(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::ProteusIdentity(self)
     }
 
     async fn find_all(
