+++ keystore/src/entities/platform/wasm/mls/e2ei_crl.rs
@@ -31,8 +31,8 @@ impl EntityBase for E2eiCrl {
         MissingKeyErrorKind::E2eiCrl
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::E2eiCrl(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::E2eiCrl(self)
     }
 
     async fn find_all(conn: &mut Self::ConnectionType, params: EntityFindParams) -> CryptoKeystoreResult<Vec<Self>> {
