+++ keystore/src/entities/platform/wasm/mls/keypackage.rs
@@ -31,8 +31,8 @@ impl EntityBase for MlsKeyPackage {
         MissingKeyErrorKind::MlsKeyPackageBundle
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::KeyPackage(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::KeyPackage(self)
     }
 
     async fn find_all(conn: &mut Self::ConnectionType, params: EntityFindParams) -> CryptoKeystoreResult<Vec<Self>> {
