+++ keystore/src/entities/platform/wasm/mls/pending_message.rs
@@ -15,8 +15,8 @@ impl EntityBase for MlsPendingMessage {
         MissingKeyErrorKind::MlsPendingMessages
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::MlsPendingMessage(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::MlsPendingMessage(self)
     }
 
     async fn find_all(conn: &mut Self::ConnectionType, params: EntityFindParams) -> CryptoKeystoreResult<Vec<Self>> {
