+++ keystore/src/entities/platform/wasm/mls/psk_bundle.rs
@@ -31,8 +31,8 @@ impl EntityBase for MlsPskBundle {
         MissingKeyErrorKind::MlsPskBundle
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::PskBundle(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::PskBundle(self)
     }
 
     async fn find_all(conn: &mut Self::ConnectionType, params: EntityFindParams) -> CryptoKeystoreResult<Vec<Self>> {
