+++ keystore/src/entities/platform/wasm/mls/refresh_token.rs
@@ -31,8 +31,8 @@ impl EntityBase for E2eiRefreshToken {
         MissingKeyErrorKind::E2eiRefreshToken
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::E2eiRefreshToken(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::E2eiRefreshToken(self)
     }
 
     async fn find_all(conn: &mut Self::ConnectionType, params: EntityFindParams) -> CryptoKeystoreResult<Vec<Self>> {
