+++ keystore/src/entities/platform/wasm/proteus/prekey.rs
@@ -32,8 +32,8 @@ impl EntityBase for ProteusPrekey {
         MissingKeyErrorKind::ProteusPrekey
     }
 
-    fn to_transaction_entity(self) -> crate::transaction::Entity {
-        crate::transaction::Entity::ProteusPrekey(self)
+    fn to_transaction_entity(self) -> crate::transaction::dynamic_dispatch::Entity {
+        crate::transaction::dynamic_dispatch::Entity::ProteusPrekey(self)
     }
 
     async fn find_all(conn: &mut Self::ConnectionType, params: EntityFindParams) -> CryptoKeystoreResult<Vec<Self>> {
